+++ selfdrive/thermald/power_monitoring.py
@@ -9,7 +9,6 @@ from common.params import Params, put_nonblocking
 from common.hardware import TICI
 from selfdrive.swaglog import cloudlog
 
-PANDA_OUTPUT_VOLTAGE = 5.28
 CAR_VOLTAGE_LOW_PASS_K = 0.091 # LPF gain for 5s tau (dt/tau / (dt/tau + 1))
 
 # A C2 uses about 1W while idling, and 30h seens like a good shutoff for most cars