+++ selfdrive/car/hyundai/interface.py
@@ -349,8 +349,8 @@ class CarInterface(CarInterfaceBase):
       events.add(EventName.belowSteerSpeed)
     if self.CC.turning_indicator_alert:
       events.add(EventName.turningIndicatorOn)
-    if not self.CS.lkas_button_on:
-      events.add(EventName.lkasButtonOff)
+    if self.CS.lkas_button_on != self.CS.prev_lkas_button:
+      events.add(EventName.buttonCancel)
     if self.mad_mode_enabled and not self.CC.longcontrol and EventName.pedalPressed in events.events:
       events.events.remove(EventName.pedalPressed)