+++ RELEASES.md
@@ -1,5 +1,6 @@
 Version 0.7.11 (2020-XX-XX)
 ========================
+ * Grey panda is no longer supported, upgrade to comma two or black panda
  * Kia Niro EV 2020 support thanks to nickn17!
 
 Version 0.7.10 (2020-10-29)